+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Mouse middle click by group tile and dashboard tile
 - Table event OnDataCompare: 
   - Ability use the token in the value
   - Ability to use regular expressions in the column name